UDLF addition to nlmeans algorithm (half completed)
The addition of the UDLF framework to the nlmeans algorithm is half
done, including creation of ranked lists based on the weight of each
pixel and change UDLF settings to use ranked lists instead of distance
matrix.

FIXME: UDLF is presenting an error where it does not run. Instead only
the message "Killed" is printed on the screen.