Get new weight indices based on the new list of weight names
The UDLF returns ranked lists in a form that each lines corresponds to
the weight/pixels names, where the earlier a name is shown in the
list, the higher must be its importance to calculate the denoised
pixel.

In this version of code, there's no function to recalculate the
weight of a pixel based on its position on the new ranked list so the
original values are still being used (This way of implementing it is
also useful to garantee this implementation is equivalent to the
original nlmeans code).