Utilization of skimage.data images for tests
If the user does not pass the image to be tested as argument, test.py
will get an image from skimage.data module, currently the image is the
camera image but this can and will be changed when needed.

Also, for comparison porpuses, the plot of the resulting image from
the subtraction im_fil1 (nlmeans) and im_fil2 (udlf_nlmeans) is being
done since in the present moment of the project, both images should
return equal (which is not happening).