Code rearrangement for the central pixel denoising
The main problem in the nlmeans_udlf function is that it's not
handling the central pixel denoising properly and that causes the
image to present black regions. The new approach follows more the
original nlmeans function where the central pixel weight is calculated
after the main loop.

At the moment IT'S NOT WORKING and has to be fixed. When trying the
most direct approach, creating a numpy array of weight and a numpy
array of names in a way that ix, iy = np.unravel_index(w_names, (M,N))
would result in ima_nse == ima_nse[ix, iy], it causes a lot of
problems I still don't know how to solve, for example, weight_indices
having shape of (443,) instead of (441,) because the new_w_names have
duplicate values. This need to be fixed.