Remove of temp code and solution found for the ranked lists denoise
The code can now correctly denoise de images based on the ranked
lists generated. The problem is that it's only true considering the
ranked lists are just a sorted version of the original arrays
w_names[i, j, :]. Since this can change after UDLF, the code probably
will needed to be refactored based on this new problem.

The solution is based on this stack overflow comment:
https://stackoverflow.com/a/69920627