Minor code refactoring
Remove of unnecessary constant RESEARCH_AREA, move code related to
UDLF config to be near the UDLF call itself. This was done for two
reasons: first, if needed to comment the UDLF part of the code for
test porpuses it's easier if all the UDLF related code is near and
second is because the L parameter of UDLF can change depending on the
type of the patch shape (for square shape it was RESEARCH_AREA but for
other patch shapes like disk it's less then RESEARCH_AREA, L being
equal to the NEIGHBOURHOOD_SIZE constant makes it correct).