Change indexation of ima_nse
When indexing using [x2range, y2range] the resulting array was
an array of shape (M,) and this was not intended. By using meshgrid
I was able to get x_idx and y_idx such that the resulting array
of the indexing ima_nse[x_idx, y_idx] was in fact a array of shape
(M, N), this array is a "translated/warped" version of the original
array and is the correct result.